wedding :D

+++ 222 PHP/Wedding/borrar.php
@@ -3,17 +3,30 @@
 
     $id = $_GET["id"];
 
-    $query = "DELETE FROM Regalos
+    $query = "SELECT * FROM Regalos
               WHERE id_regalo = $id";
 
-    if(mysqli_query($conexion, $query))
+    $row = mysqli_fetch_assoc(mysqli_query($conexion, $query));
+
+    if(unlink($row["imagen"]))
     {
-        echo "Registro Borrado";
-        header("Location:lista.php");
+        $query = "DELETE FROM Regalos
+                  WHERE id_regalo = $id";
+
+        if(mysqli_query($conexion, $query))
+        {
+            echo "Registro Borrado";
+            header("Location:lista.php");
+        }
+        else
+        {
+            echo "Error: ".mysqli_error($conexion);
+        }
     }
     else
     {
-        echo "Error: ".mysqli_error($conexion);
+        echo "Error: No se pudo borrar la imagen.";
     }
+    
     mysqli_close($conexion);
 ?>
